+++ controlplane/src/core/bufservices/organization/whoAmI.ts
@@ -29,6 +29,7 @@ export function whoAmI(
         },
         organizationName: '',
         organizationSlug: '',
+        organizationId: '',
       };
     }
 
@@ -36,6 +37,8 @@ export function whoAmI(
       response: {
         code: EnumStatusCode.OK,
       },
+      organizationId: organization.id,
+      userEmail: authContext.userDisplayName,
       organizationName: organization.name,
       organizationSlug: organization.slug,
     };